fix: add support for firebase hosting proxy
The proxy between firebase hosting and firebase functions do not “x-formarded-proto” header. It uses a app engine header (“x-appengine-https” set to “on”) and fastly header (“fastly-ff” set to “1”).

I am not sure which one is in front. I am assuming it’s the app engine one and using it check if the protocol of the original request protocol is HTTPS.