chore: remove “lti” prefix from keys
Firebase keys start with “-”. The key was provided with a “lti” prefix so that it would start with an alphanumeric character.

It’s error prone since it’s easy to forget to add back the key for validating/signing oauth requests.

Using a uuid key instead.

BREAKING CHANGE: keys need to have the “lti” prefix in the db or need to be recreated.